fix: fix usage of alignments
Previously, the CLI transformation wasn't configured in a way that made sure
that the loaded source files are loaded into the correct types of the source
schema. This led to the situation that the generated GeoPackage output files
were incomplete.

This extends the configuration so that the target type for each Shapefile can
be provided. Also, the import of the Excel tables can be configured in more
detail to allow reading multiple sheets.

SVC-1216