refactor(df7_10): add explicit `-plan` tasks
Previously, all validation and aggregation tasks for plans depended directly on
the corresponding task for coverage areas. That led to the situation that plan
tasks were not executed in case the corresponding coverage area task failed.

With this change, plan validation and aggregation tasks are renamed to have a
`-plan` suffix. A new meta task without a suffix is added that executes both
plan and coverage areas tasks.

SVC-1812