fix(df7_10): add separate task for Agglomerations Quiet Areas
As the `Ruhige_Gebiete` sheet has a different index in the Excel workbook
template for Agglomerations than in the templates for other plan types, a
different configuration of the Excel reader is required for Agglomerations.

This adds a separate task for deriving Quiet Areas from Agglomeration plans
using a separate input folder.

SVC-1812